filter data models separated from others, cart and seller data models fixed, some filter UI fixes

+++ app/src/main/java/ist/cabin/cabinCustomerBase/models/backend/JSONSeller.kt
@@ -7,6 +7,6 @@ import com.squareup.moshi.JsonClass
 data class JSONSeller(@Json(name = "ID") val id: Int,
                       @Json(name = "NAME") val name: String,
                       @Json(name = "PRODUCT") val products: List<JSONProduct?>,
-                      @Json(name = "SELLERSHIPPINGPRICE") val shippingPrice: Int?,
-                      @Json(name = "SELLERSUBTOTAL") val subtotal: Int?,
-                      @Json(name = "SELLERTOTAL") val total: Int)
+                      @Json(name = "SELLERSHIPPINGPRICE") val shippingPrice: Double?,
+                      @Json(name = "SELLERSUBTOTAL") val subtotal: Double?,
+                      @Json(name = "SELLERTOTAL") val total: Double)
